Add local dist meta data persist repository properties item
