+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/config/persist/repository/local/LocalRepositoryProperties.java
@@ -15,41 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.config.persist.repository;
+package org.apache.shardingsphere.infra.config.persist.repository.local;
 
-import java.util.Collections;
-import java.util.List;
+import org.apache.shardingsphere.infra.properties.TypedProperties;
+import java.util.Properties;
 
 /**
- * Local dist meta data persist repository.
+ * Local repository properties.
  */
-// TODO finish me
-public final class LocalDistMetaDataPersistRepository implements DistMetaDataPersistRepository {
+public final class LocalRepositoryProperties extends TypedProperties<LocalRepositoryPropertyKey> {
     
-    @Override
-    public String get(final String key) {
-        return "";
-    }
-    
-    @Override
-    public List<String> getChildrenKeys(final String key) {
-        return Collections.emptyList();
-    }
-    
-    @Override
-    public void persist(final String key, final String value) {
-    }
-    
-    @Override
-    public void delete(final String key) {
-    }
-    
-    @Override
-    public void close() {
-    }
-    
-    @Override
-    public String getType() {
-        return "Local";
+    public LocalRepositoryProperties(final Properties props) {
+        super(LocalRepositoryPropertyKey.class, props);
     }
 }
